Backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Bybit Trading
pybit==5.3.0  # SDK oficial de Bybit
ccxt==4.1.59  # Como alternativa Puede ser útil como alternativa o para multi-exchange en el futuro

# Data Analysis
pandas==2.1.3
numpy==1.25.2
ta==0.10.2 # Análisis técnico

# Notifications
python-telegram-bot==20.7
requests==2.31.0

# Utilities
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1

# WebSocket para datos en tiempo real
websockets==12.0
websocket-client==1.6.3